// src/generated/api/models/LeaderboardEntry.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LeaderboardEntry = {
    /**
     * Participant ID
     */
    participantId: number;
    /**
     * Participant nickname
     */
    nickname: string;
    /**
     * Participant avatar
     */
    avatar: string;
    /**
     * User ID
     */
    userId: number;
    /**
     * Total achievement points
     */
    achievementPoints: number;
    /**
     * Total medal points
     */
    medalPoints: number;
    /**
     * Total points (achievements + medals)
     */
    totalPoints: number;
    /**
     * Number of achievements completed
     */
    achievementCount: number;
    /**
     * Number of medals earned
     */
    medalCount: number;
    /**
     * Current rank position
     */
    rank: number;
};